parse url and check for root pathname
